requirements.txt
#############################################
# 기본 파이썬 라이브러리
#############################################
numpy==1.26.4
pandas
tqdm
matplotlib
seaborn
pillow
opencv-python

#############################################
# 파일/경로 관련
#############################################
pathlib
pyyaml

#############################################
# 머신러닝 / 딥러닝 (CUDA 12.1 호환)
#############################################
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

#############################################
# CLIP(OpenCLIP)
#############################################
open_clip_torch

#############################################
# Ultralytics YOLOv8 (공식 안정 버전)
#############################################
ultralytics==8.1.0

#############################################
# 기타
#############################################
scikit-learn